Value attribute in PQV,  REAL should be of type Double instead of BigDecimal to be inconsistent with ISO xsd.
SVN-Revision: 206

+++ localization/src/test/java/gov/nih/nci/iso21090/RealTest.java
@@ -82,8 +82,6 @@
  */
 package gov.nih.nci.iso21090;
 
-import java.math.BigDecimal;
-
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -133,7 +131,7 @@ public class RealTest {
 
            assertEquals(first.hashCode(), second.hashCode());
 
-           second.setValue(new BigDecimal(555.0));
+           second.setValue(new Double(555.0));
 
            assertFalse(first.hashCode() == second.hashCode());
 
@@ -143,7 +141,7 @@ public class RealTest {
        public void testCloneable() {
            Real uncertainty1 = new Real();
            uncertainty1.setNullFlavor(NullFlavor.DER);
-           uncertainty1.setValue(new BigDecimal(111.0));
+           uncertainty1.setValue(new Double(111.0));
 
            Real first = new Real();
            first.setNullFlavor(NullFlavor.ASKU);
